Change default target behavior to 'keep at range' instead of 'orbit'
This is due to a rare occurrence in which the asteroid currently being mined
gets destroyed but no popup appears. If the ship is orbiting the asteroid while
this happens, it will continue to fly straight in a random direction until the
script times out. If, instead, the script uses 'keep at range', the ship will
just sit idle, simulating an AFK user.

Signed-off-by: takelley1 <37026149+Zultair@users.noreply.github.com>