Add an 'unsuitable_site' counter to prevent endlessly searching for a mining site
This is a global variable that increases every time the miner script finds
a site without any asteroids or with undesirable npcs/pcs. This is to prevent
the script from, in the rare case that all sites are empty, constantly
warping around from site to site forever. This variable resets to 0 when
the ship locates an asteroid. Once the limit is reached, the ship will dock if
possible and the script will terminate.

Signed-off-by: takelley1 <37026149+Zultair@users.noreply.github.com>