Add sleep timer for miner to wait before starting mining lasers
This is an experimental sleep timer to prevent the script from constantly
attempting to activate its mining lasers when it's locked a target but its
modules are still out of range. The purpose of the sleep is to wait for the
ship to get closer to its target so it will be in range when it attempts to
start its mining lasers.

Signed-off-by: takelley1 <37026149+Zultair@users.noreply.github.com>